add função
adicionei a função de inserirDados

+++ include/data.hpp
@@ -14,7 +14,7 @@ private:
     int average;
 
 public :
-    Data( int registration, string name, string course, int grade1, int grade2, int grade3, int average );
+    Data( float registration, string name, string course, float grade1, float grade2, float grade3, float average );
     ~Data();
 
 };
