+++ src/data.cpp
@@ -7,7 +7,7 @@
 using namespace std;
   
 //Construtor
-Data::Data(int registration, string name, string course, int grade1, int grade2, int grade3, int average){   
+Data::Data(float registration, string name, string course, float grade1, float grade2, float grade3, float average){   
     this->registration = registration;
     this->name = name;
     this->course = course;
@@ -19,7 +19,7 @@ Data::Data(int registration, string name, string course, int grade1, int grade2,
 
 //Destrutor
 Data::~Data(){
-    
+
 }
